==> NBA_Data/Part1_NBA_Draft_History.ipynb <==
# What's this about?
For the next few blog posts I'll be dealing with NBA draft data. If you pay attention to the NBA at all, then you'll be aware of what the draft is and it's presumed importance. Basketball analysts such as Zach Lowe and Bill Simmons have  argued that the best chance a bad team has at being a good team is to have a high draft pick, while many empirical studies show the value of draft position: [here][3], [here] [2], and [here] [1] and general managers have staked their careers on the premise that the way to a great team was through drafting a star player. Given the importance of the draft then, I decided to take a look at the link between draft position and player quality (pretty well established correlation) as well as looking at whether certain GM's or coaches consistently beat the odds in the draft. 
Since teams have equivalent information on the players they are going to draft, there should be a clear link between the quality of the player and the player's draft position. GM's and coaches that consistently get good players from lower draft positions then either are evaulating the available information differently or better at developing players. Teasing these two apart may not be possible, but given that Sam Hinkie based his career on having high draft picks, I think it's worth looking at whether there are GM's or coaches that are able to grab players that exceed the average player's performance given the same draft position. This analysis has actually been done [before][4] (and [again][5]) so my analysis will be extending this to more years as well as looking to pull apart the influence of the coach as compared to the GM in drafting performance. Another advantage of this project, despite being done before, is that I have something to compare my results too (granted I'll need to find access to these projects) and to look to for ideas on how to actually analyze the data.
# What does this project teach me?
To complete this project, I'll need to scrape data from a number of websites, convert them into Pandas DataFrames and merge them. Most of this project will be gathering and cleaning data, but eventually I will be graphing player performance based against draft position and analyzing which GM's and coaches did better in the draft.

[1]: http://nyloncalculus.com/2016/03/07/evaluating-draft-prospects-a-first-pass/
[2]: http://www.82games.com/barzilai1.htm
[3]: https://statsbylopez.com/2016/06/22/the-making-and-comparison-of-draft-curves/Teams
[4]: http://basketball.realgm.com/wiretap/232776/study-finds-isiah-thomas-as-best-drafting-gm-since-1989
[5]: http://www.nbaminer.com/draft-success-of-nba-teams/

# And it begins...
This project will be carried out in parts and I will be trying to share how I did things as well as the difficulties I encountered. To begin, I am trying to grab draft data by following this [blog post][1], but as you'll see, I quickly deviated because I ran into many errors and tried whatever I could to get around these issues. Part 1 of this project will be focused on getting the draft data from a single season, 1976 (the NBA merger season) with Part 2 focused on using the template here to create a loop to grab all draft data from 1976 to 2014 and combine them into a single large DataFrame.
[1]: http://savvastjortjoglou.com/nba-draft-part01-scraping.html

import pandas as pd
from pandas import Series,DataFrame,read_html
import numpy as np

from bs4 import BeautifulSoup
import html5lib

Above is some standard imports. I tried to follow the blog, but despite have installed requests, kept getting an error that the module was not available, so I switched tactics after finding pandas 'read_html' command. 
Below I will go through scraping data from a single season and then progressing to build a loop to grab all the seasons that I want before finally merging all the DataFrames together.

url = 'http://www.basketball-reference.com/draft/NBA_1976.html'

dframe_list = pd.io.html.read_html(url)

type(dframe_list)

You can convert the list to a Pandas DataFrame by simply assigning the first element in the list. I'm not quite sure how Pandas knows that I want to create a DataFrame, but it takes the first element in my list (which is a list of column headers) and makes them the column headers in a DataFrame.

dframe = dframe_list[0]
dframe.head()

type(dframe)

dframe.columns.values

Above I've shown that dframe is a DataFrame and printed out my column headers. As you can see, these column headers are clearly wrong. The blog I was following deals with this issue by taking a copy of the website and making it a beautiful soup object, then searching through that object for the column headers and then assigning those as the column headers. As I'll briefly discuss below, for some reason that did not work for me, instead, I continually got an empty list when I tried to grab the column headers and took an alternative route that may not be ideal for a data set with a vast number of column headers (so sooner or later I'll need to learn the first method). 

First, in looking at the dataset, I noticed that the dataframe had been read in to have the first and second rows from the webpage to be the column headers (look at the webpage to see what I mean). This resulted in misalignment of the values and a bunch of columns at the end filled with NaN values. My next steps then are to drop those NaN filled columns and to create a list of accurate column headers to replace with the misaligned column headers. Finally, because the webpage has rows that provide breaks for different rounds of the draft, I need to drop these useless rows. 

# dropping the NaN filled columns
# dframe.dropna(axis=1,how='all',inplace=True)

I tried the above code first to drop those NaN filled columns, which worked, but also dropped a bunch of rows of players that never played in the NBA and thus all of their column entries were NaN at the beginning. Since I know there are 22 actual columns, I will drop all columns after the 22nd.

dframe.drop(dframe.columns[[22,23,24,25,26,27,28,29,30]],inplace=True,axis=1)
# For some reason, trying to enter a range of number e.g. 22:30 for the dropped columns returned a syntax error

# the list of accurate column headers, this is pretty inefficient to just create the list
column_names = ['Rk','Pk','Team','Player','College','Yrs','Games','Minutes Played','PTS','TRB','AST','FG_Percentage','TP_Percentage','FT_Percentage','Minutes per Game','Points per Game','TRB per game','Assits per Game','Win Share','WS_per_game','BPM','VORP']

Now I can replace the column headers with the list of column headers I created here. As you can tell, I'm having to do some of these things by plugging in values and such which is not ideal, so I still have a ways to go in learning more the syntax of Python and pandas.

# replacing the column headers with the properly named list created above
dframe.columns=column_names

dframe.head()

## An Aside
In trying to follow the same steps as the blog to replace column headers, I kept running into an error every time I tried to replace the column headers with those that I found using 'soup.finall' that I did not have a [1] list item. I ended up assigning the command to an object 'rows' and in looking at 'rows', as shown below, found it to be an empty list. I poked around quite a bit online and there are lots of people that seem to run into this issue, but for a wide variety of reasons. At this point I'm not sure if it's something to do with the versions of python and beautifulsoup that I'm running or something else. For the sake of simplicity I decided to simply use the method demonstrated above and created a list of column_names.

rows = soup.findAll('tr', limit=2)
rows

There are a couple of things I need to do now. 
1. I need to drop the rows that served as breaks for different rounds in the original dataset
2. I'll need to re-align the index of the rows to be correct
3. I need to fill the NaN values in as zeroes
4. I need to fix the object type for each column
Since having the correct object type for each column can speed up the the other three things I need to do, I'll do that first.

dframe.dtypes

# Since the Team, Player, College columns do not need to be converted I will leave these out
numeric_columns = column_names
del numeric_columns[2:5]

for column in numeric_columns:
    dframe[column] = pd.to_numeric(dframe[column], errors='coerce')
dframe.dtypes

Now that I have these in numeric form, I should be able to drop the excess rows and then convert the remaining NaN's to zeroes


# dropping the rows that served as breaks for different rounds of the draft
dframe = dframe[dframe.Rk.notnull()]
dframe.head()

# changing the remaining NaN's to zeroes
dframe = dframe.fillna(0)
# reindexing to align the row index correctly
dframe.index = range(173)
dframe.head()

# checking whether there are still any missing values
dframe.isnull().sum().sum()

## Moving forward
Now that I have the dataframe with proper column headers and all rows are properly filled in, I need to add a column for the draft year and can drop the 'Rk' column.

# Inserting a column for draft year 
dframe.insert(0, 'Draft_Yr', 1976)

# Deleting the 'Rk' column
dframe.drop('Rk', axis=1, inplace=True)
dframe.head()

cd NBA_Data

dframe.to_csv('1976_Draft.csv')

# And I finally have it...
Creating this first DataFrame for one year of the draft took quite a bit more time than I would have thought (about a week working for a few hours in the evening when there was time.) Even with a blog post to follow, I ran into quite a few errors that required searching around the internet (mainly stackexchange) to find solutions. Python is trickier in its syntax than what I'm used to with Stata and whereas in Stata I would have solved some of these problems by building loops, Python has ready made commands to solve the problems while it's loops are still a little more difficult for me to implement. Now on to the second part of the blog post, creating a loop to grab all this draft data and putting it into a single DataFrame.
