==> NBA_Data/Part5_NBA_Draft_History.ipynb <==
# Some data cleaning
So I was able to create the dataset I desired with GM data. I still would like to add coaching data, but with what I have I can start to run some analysis. I've been lucky enough to bounce analysis ideas off of a couple of friends (CJ and Adam thank you) and who have been very helpful running a little preliminary analysis to see if what comes out makes any sense and checking for errors in the data. This led to this hopefully short post on some data cleaning and getting it ready for analysis.

## Inputting Trades
In my last post I ended by stating that I wanted to input draftee trade deals such that the GM that traded for the draftee would be credited with the pick. I wanted to do this for two reasons: (1) I don't believe that a team like Charlotte should be credited with a pick like Kobe Bryant when they traded him away before seeing his true value (or conversely if he had been terrible, then the Lakers should have been 'blamed' for that pick) and (2) teams sometimes draft a player with a trade already in mind and the pick is made on behalf of another team. In trying to put these trade deals in, I was having to do it by hand, as all the information was not contained in one site. Unfortunately, this had many drawbacks, namely that I couldn't verify whether I was getting all the trades. It also seems that BR had put in some of these trade deals, but not others. Whether BR has inputted them or not, I'm not sure what the criteria are (perhaps the trade had to happen prior to a certain point on draft day?). Regardless, since I couldn't easily verify trades and that I realized if I was instituting these deals then really I should be assigning to the player the team that he signed with his rookie season (which might also be different from the team that drafted him if the drafting team waives their rights to sign him) all led me to abandon this editing of the data. Though I still consider this to be the more appropriate data, for my purposes of learning Python and demonstrating that I know how to analyze data, I will simply use the BR data as it came. 
I am stating all this up front to also demonstrate that I understand issues that arise with data and am able to address them. Now, onto the cleaning and manipulating.

# Standard imports 
import pandas as pd
from pandas import Series,DataFrame,read_html
import numpy as np

# Loading up the dataframe
df = pd.read_csv('NBA_Data/1976_to_2015_Draftees.csv')

For use in later analysis on how well a GM drafted, I believe it would be useful to know if the player attended college. Ideally I would institute the total number of years a player played in college, but for this analysis I'm trying to go a little more quickly to get to the analysis and don't have as much time to find all the college basketball info. 
The reason I'm adding whether a player went to college as a dummy variable is that a player with college experience provides more information to a GM when drafting. You would likely expect that players who have played in college are better known quantities and thus GM's should have less variance in drafting such players. Unfortunately, this variable only captures if a player played in college, so it misses those that played overseas before coming to the US. Players coming from foreign leagues also might offer more data insight prior to draft, but in this analysis will be grouped with those coming right out of high school which we would expect to close the gap some between the variance in player quality when drafted between college and non-college drafted players. Ultimately this is not a significant problem though as it should simply lead to underestimating the impact of having gone to college.

# Creating a dummy variable for if a player went to college or not
df['attend_college'] = 1
df.loc[df['College'] == '0', 'attend_college'] = 0

Wanted to show that I know how to make a function, even if it's a really simple one!

# Alternative method for creating the dummy variable 
def dummyVar(df):
    if df['College'] == '0':
        return 0
    else:
        return 1

df['attend_college2'] = df.apply(dummyVar, axis=1)

# Demonstrating that those players who did not go to college also got a zero in the 'attend_college' column
df[df['attend_college'] == 0].head()

# Demonstrating that those players who did  go to college also got a one in the 'attend_college' column
df[df['attend_college'] == 1].head()

# This bit of code creates a unique executive id and assigns it to each row of the executive
# While this isn't necessary for my analysis, it may come in handy later and definitely would be 
# necessary were I dealing with a larger dataset where speed of analysis was an issue.

keys = pd.unique(df.Executive.ravel()) # returns an array of each unique executive
d = dict(zip(keys, range(1,len(keys)))) # create a dictionary of unique id's for each executive
df['Exec_id'] = df.Executive.map(d) # mapping executive id's to the executive in the dataframe

df = df.sort_values(by=['Executive', 'Exec_id']) 
df.ffill() # giving each instance of an executive the proper exec_id
df.head()

## GM Drafting Experience
I next wanted to create a column denoting for each pick how many years of experience the GM had at drafting at that point. So for a GM's first year of drafting this column would be one, in his second year this would be two, and so on. Though this is likely to be heavily correlated with draft pick quality (you'd expect GM's who drafted well get to keep on drafting), I think it would be good to control for this as we would also expect GM's with more experience in drafting to draft better in later years. This seems particularly interesting for GM's like Larry Bird who have come in and out of the position. Is Larry Bird as good at drafting coming back to the job say four years after his last draft, or is it likely that a GM loses touch with the game to the extent that what one looks for in a player has changed?

To put in this column, at first I thought to write a function again, similar to the one above. I almost immediately realized that this would be more difficult than the Python method 'Categorical' and that my function variable 'count' was not going to work. I left it here anyway to show a little of my thought process and that I still screw up frequently.

# failed function
def draftExp2(df):
    count = 1
    if df['Exec_draft_exp2'] == 0:
        count = count + 1
        return count
    elif df['Exec_draft_exp2'] != 0:
        return 1
    
df['Exec_draft_exp2'] = df.apply(draftExp2, axis=1)
df

# Very easy method to use from the pandas library
df['Exec_draft_exp'] = pd.Categorical(pd.factorize(df.Draft_Yr)[0] + 1)

# Demonstrating the new Exec_draft_exp column!
df[df['Executive'] == 'Al Attles'].head(25)

## Duplicated players
Now comes some quality checks on potential duplicates. I found a number of them and there are two types of duplicates.
Type 1: Players who were drafted, but then decided to not enter the NBA and went back to college (see Mark Eaton) and then were later drafted again. These players have exactly the same stats, but different draft years, pick number, team, and GM's. I will drop the earlier entry, though I think an argument can be made that these players should simply be left in the dataframe. I am choosing to drop them because playing more basketball before entering the NBA provided the later GM with more information to assess player ability and thus that GM with greater information should have made a better choice than the earlier GM. But the earlier GM should also not get 'credit' for a player that took more time to develop in college and potentially entered the NBA at the 'right time' for himself.
Type 2: Players who have the same name as another player. There are more of these than one would think and they present a challenge as when I merged the player draft data with the player accomplishment data, both players picked up the player accomplishment data (all-star appearances and all-NBA teams). I deal with these later.

# Type 1 players, there are 24 in all
df[df.duplicated(['Player','Yrs','Games','College'])]

# Dropping the first instance of the Type 1 duplicates
df.drop_duplicates(subset = ['Player', 'College', 'Yrs', 'Games'], keep = 'last', inplace=True)

# Showing that there are no more Type 1 duplicates
df[df.duplicated(['Player','Yrs','Games','College'])]

## Check to make sure it took the later entry
Aryvdas Sabonis was a type 1 duplicate. I am loading his information here now to show that I did in fact drop the earlier entry and kept the 'duplicate'.

df[df['Player'] == 'Arvydas Sabonis']

# The Type 2 duplicates, there are 56 in all
df[df.duplicated(['Player'])].head()

## Only a small problem for the merged player accomplishment data though
In looking through thte Type 2 duplicates I can look at just the All-NBA column to see where a mistake with merging the player accomplishment data might have arose. Since there are only 55 Type 2 duplicates, this is fairly quick and I found only 4 where this occurred (Patrick Ewing, Glen Rice, Tim Hardaway, Larry Johnson). Given how few there are I will just fix this by hand in the final dataframe.

# I hate unneeded columns, so here I am looking at the columns and seeing which can be dropped
df.columns

# Dropping columns
df.drop(df.columns[[0,-3]], inplace=True, axis=1)

cd NBA_Data

df.to_csv('1976_to_2015_Draftees_edit.csv')

# Finally, next some analysis
So I've cleaned the data and added some columns I'll use later for analysis, but I'm really anxious to get to analyzing. My next post I'll actually be doing some stuff to pull out insights from the data. I'll actually be doing this with some help from a friend and I'll be using R for the analysis. I've been playing around with learning R recently and it seems easier and more friendly for actual analysis work. The friend I will be collaborating with also already poked around the data using R, so we will be building off some of his preliminary work.
I will also be trying to carry out the same analysis in Python, but I will be putting those only in a notebook that I share on github and not as a blog post.
